+++ frontend/src/features/permits/pages/Application/ApplicationReview.tsx
@@ -23,22 +23,54 @@ import { DEFAULT_PERMIT_TYPE, PERMIT_TYPES } from "../../types/PermitType";
 import { PERMIT_REVIEW_CONTEXTS } from "../../types/PermitReviewContext";
 import { usePolicyEngine } from "../../../policy/hooks/usePolicyEngine";
 import { useCommodityOptions } from "../../hooks/useCommodityOptions";
-import { useSubmitApplicationForReview } from "../../../queue/hooks/hooks";
 import { deserializeApplicationResponse } from "../../helpers/serialize/deserializeApplication";
 import OnRouteBCContext from "../../../../common/authentication/OnRouteBCContext";
 import {
   APPLICATIONS_ROUTES,
+  APPLICATION_QUEUE_ROUTES,
   APPLICATION_STEPS,
+  APPLICATION_STEP_CONTEXTS,
+  ApplicationStepContext,
   ERROR_ROUTES,
+  IDIR_ROUTES,
 } from "../../../../routes/constants";
+import { CASE_ACTIVITY_TYPES } from "../../../queue/types/CaseActivityType";
+import { QueueBreadcrumb } from "../../../queue/components/QueueBreadcrumb";
+import { RejectApplicationModal } from "../../../queue/components/RejectApplicationModal";
+import {
+  useSubmitApplicationForReview,
+  useUpdateApplicationInQueueStatus,
+} from "../../../queue/hooks/hooks";
+import { Nullable } from "../../../../common/types/common";
+import { UnavailableApplicationModal } from "../../../queue/components/UnavailableApplicationModal";
 
 export const ApplicationReview = ({
-  companyId,
+  companyIdProp,
+  applicationStepContext,
+  applicationData: queueApplicationData,
 }: {
-  companyId: number;
+  applicationStepContext: ApplicationStepContext;
+  companyIdProp?: Nullable<number>;
+  applicationData?: Nullable<Application>;
 }) => {
   const { applicationData, setApplicationData: setApplicationContextData } =
     useContext(ApplicationContext);
+
+  const isQueueContext =
+    applicationStepContext === APPLICATION_STEP_CONTEXTS.QUEUE;
+
+  const application = isQueueContext ? queueApplicationData : applicationData;
+
+  const companyId = getDefaultRequiredVal(
+    0,
+    companyIdProp,
+    queueApplicationData?.companyId,
+  );
+
+  const applicationId = getDefaultRequiredVal(
+    queueApplicationData?.permitId,
+    applicationData?.permitId,
+  );
 
   const { idirUserDetails } = useContext(OnRouteBCContext);
   const isStaffUser = Boolean(idirUserDetails?.userRole);
@@ -49,12 +81,15 @@ export const ApplicationReview = ({
   const { data: companyInfo } = useCompanyInfoDetailsQuery(companyId);
   const doingBusinessAs = companyInfo?.alternateName;
 
-  const permitType = getDefaultRequiredVal(DEFAULT_PERMIT_TYPE, applicationData?.permitType);
+  const permitType = getDefaultRequiredVal(
+    DEFAULT_PERMIT_TYPE,
+    application?.permitType,
+  );
   const fee = isNoFeePermitType
     ? "0"
     : `${calculateFeeByDuration(
         permitType,
-        getDefaultRequiredVal(0, applicationData?.permitData?.permitDuration),
+        getDefaultRequiredVal(0, application?.permitData?.permitDuration),
       )}`;
 
   const { setSnackBar } = useContext(SnackBarContext);
@@ -71,20 +106,25 @@ export const ApplicationReview = ({
   const trailerSubTypesQuery = useTrailerSubTypesQuery();
   const methods = useForm<Application>();
 
-  // For the confirmation checkboxes
-  const [allConfirmed, setAllConfirmed] = useState(false);
+  const [allConfirmed, setAllConfirmed] = useState(isQueueContext);
   const [hasAttemptedSubmission, setHasAttemptedSubmission] = useState(false);
 
   const { mutateAsync: saveApplication } = useSaveApplicationMutation();
   const addToCartMutation = useAddToCart();
-
-  // Submit for review (if applicable)
+  const { mutateAsync: submitForReview } = useSubmitApplicationForReview();
   const {
-    mutateAsync: submitForReview,
-  } = useSubmitApplicationForReview();
+    mutateAsync: updateApplication,
+    data: updateApplicationResponse,
+    isPending: updateApplicationMutationPending,
+  } = useUpdateApplicationInQueueStatus();
 
-  const back = () => {
-    navigate(APPLICATIONS_ROUTES.DETAILS(permitId), { replace: true });
+  const handleEdit = () => {
+    navigate(
+      isQueueContext
+        ? APPLICATION_QUEUE_ROUTES.EDIT(companyId, applicationId)
+        : APPLICATIONS_ROUTES.DETAILS(permitId),
+      { replace: true },
+    );
   };
 
   const handleSaveApplication = async (
@@ -98,42 +138,45 @@ export const ApplicationReview = ({
 
     if (!allConfirmed) return;
 
-    const companyId = applicationData?.companyId;
-    const permitId = applicationData?.permitId;
-    const applicationNumber = applicationData?.applicationNumber;
+    const companyId = application?.companyId;
+    const permitId = application?.permitId;
+    const applicationNumber = application?.applicationNumber;
     if (!companyId || !permitId || !applicationNumber) {
       return navigate(ERROR_ROUTES.UNEXPECTED);
     }
 
-    await saveApplication({
-      data: {
-        ...applicationData,
-        permitData: {
-          ...applicationData.permitData,
-          doingBusinessAs, // always set most recent DBA from company info
+    await saveApplication(
+      {
+        data: {
+          ...application,
+          permitData: {
+            ...application?.permitData,
+            doingBusinessAs,
+          },
+        },
+        companyId,
+      },
+      {
+        onSuccess: ({ data: savedApplication }) => {
+          setApplicationContextData(
+            deserializeApplicationResponse(savedApplication),
+          );
+          followUpAction(companyId, permitId, applicationNumber);
+        },
+        onError: (e) => {
+          console.error(e);
+          if (isAxiosError(e)) {
+            navigate(ERROR_ROUTES.UNEXPECTED, {
+              state: {
+                correlationId: e?.response?.headers["x-correlation-id"],
+              },
+            });
+          } else {
+            navigate(ERROR_ROUTES.UNEXPECTED);
+          }
         },
       },
-      companyId,
-    }, {
-      onSuccess: ({ data: savedApplication }) => {
-        setApplicationContextData(
-          deserializeApplicationResponse(savedApplication),
-        );
-        followUpAction(companyId, permitId, applicationNumber);
-      },
-      onError: (e) => {
-        console.error(e);
-        if (isAxiosError(e)) {
-          navigate(ERROR_ROUTES.UNEXPECTED, {
-            state: {
-              correlationId: e?.response?.headers["x-correlation-id"],
-            },
-          });
-        } else {
-          navigate(ERROR_ROUTES.UNEXPECTED);
-        }
-      },
-    });
+    );
   };
 
   const proceedWithAddToCart = async (
@@ -156,49 +199,93 @@ export const ApplicationReview = ({
   const setShowSnackbar = () => true;
 
   const handleAddToCart = async () => {
-    await handleSaveApplication(async (companyId, permitId, applicationNumber) => {
-      await proceedWithAddToCart(companyId, [permitId], () => {
-        setSnackBar({
-          showSnackbar: true,
-          setShowSnackbar,
-          message: `Application ${applicationNumber} added to cart`,
-          alertType: "success",
-        });
-
-        refetchCartCount();
-        navigate(APPLICATIONS_ROUTES.BASE);
-      });
-    });
-  };
-
-  const continueBtnText = permitType === PERMIT_TYPES.STOS && !isStaffUser
-    ? "Submit for Review" : undefined;
-
-  const handleSubmitForReview = async () => {
-    if (permitType !== PERMIT_TYPES.STOS) return;
-    if (isStaffUser) return;
-
-    await handleSaveApplication(async (companyId, permitId, applicationNumber) => {
-      await submitForReview({
-        companyId,
-        applicationId: permitId,
-      }, {
-        onSuccess: () => {
+    await handleSaveApplication(
+      async (companyId, permitId, applicationNumber) => {
+        await proceedWithAddToCart(companyId, [permitId], () => {
           setSnackBar({
             showSnackbar: true,
             setShowSnackbar,
-            message: `Application ${applicationNumber} submitted for review`,
+            message: `Application ${applicationNumber} added to cart`,
             alertType: "success",
           });
-  
+
+          refetchCartCount();
           navigate(APPLICATIONS_ROUTES.BASE);
-        },
-        onError: () => {
-          navigate(ERROR_ROUTES.UNEXPECTED);
-        },
-      });
+        });
+      },
+    );
+  };
+  const continueBtnText =
+    permitType === PERMIT_TYPES.STOS && !isStaffUser
+      ? "Submit for Review"
+      : undefined;
+
+  const handleSubmitForReview = async () => {
+    if (permitType !== PERMIT_TYPES.STOS || isStaffUser) return;
+
+    await handleSaveApplication(
+      async (companyId, permitId, applicationNumber) => {
+        await submitForReview({ companyId, applicationId: permitId });
+        setSnackBar({
+          showSnackbar: true,
+          setShowSnackbar: () => true,
+          message: `Application ${applicationNumber} submitted for review`,
+          alertType: "success",
+        });
+        navigate(APPLICATIONS_ROUTES.BASE);
+      },
+    );
+  };
+
+  const handleApprove = async () => {
+    setHasAttemptedSubmission(true);
+    await updateApplication({
+      applicationId: permitId,
+      companyId,
+      caseActivityType: CASE_ACTIVITY_TYPES.APPROVED,
     });
   };
+
+  const [showRejectApplicationModal, setShowRejectApplicationModal] =
+    useState<boolean>(false);
+
+  const handleRejectButton = () => {
+    setShowRejectApplicationModal(true);
+  };
+
+  const handleReject = async (comment: string) => {
+    setHasAttemptedSubmission(true);
+    await updateApplication({
+      applicationId: permitId,
+      companyId,
+      caseActivityType: CASE_ACTIVITY_TYPES.REJECTED,
+      comment,
+    });
+  };
+
+  const [showUnavailableApplicationModal, setShowUnavailableApplicationModal] =
+    useState<boolean>(false);
+
+  const [currentClaimant, setCurrentClaimant] = useState<string>("");
+
+  const updateApplicationResponseStatus = updateApplicationResponse?.status;
+
+  const handleCloseApplication = () => {
+    navigate(IDIR_ROUTES.STAFF_HOME);
+  };
+
+  useEffect(() => {
+    if (updateApplicationResponseStatus === 201) {
+      navigate(IDIR_ROUTES.STAFF_HOME);
+    }
+    if (updateApplicationResponseStatus === 422) {
+      // TODO since I do not have access to two separate IDIR accounts, this will need to be checked in test environment
+      setCurrentClaimant(
+        updateApplicationResponse?.data.error[0].additionalInfo.currentClaimant,
+      );
+      setShowUnavailableApplicationModal(true);
+    }
+  }, [updateApplicationResponse, navigate]);
 
   useEffect(() => {
     window.scrollTo(0, 0);
@@ -206,14 +293,25 @@ export const ApplicationReview = ({
 
   return (
     <div className="application-review">
-      <ApplicationBreadcrumb
-        permitId={permitId}
-        applicationStep={APPLICATION_STEPS.REVIEW}
-      />
+      {isQueueContext ? (
+        <QueueBreadcrumb
+          applicationNumber={applicationData?.applicationNumber}
+          applicationStep={APPLICATION_STEPS.REVIEW}
+        />
+      ) : (
+        <ApplicationBreadcrumb
+          permitId={permitId}
+          applicationStep={APPLICATION_STEPS.REVIEW}
+        />
+      )}
 
       <FormProvider {...methods}>
         <PermitReview
-          reviewContext={PERMIT_REVIEW_CONTEXTS.APPLY}
+          reviewContext={
+            isQueueContext
+              ? PERMIT_REVIEW_CONTEXTS.QUEUE
+              : PERMIT_REVIEW_CONTEXTS.APPLY
+          }
           permitType={permitType}
           permitNumber={applicationData?.permitNumber}
           applicationNumber={applicationData?.applicationNumber}
@@ -228,10 +326,23 @@ export const ApplicationReview = ({
           updatedDateTime={applicationData?.updatedDateTime}
           companyInfo={companyInfo}
           contactDetails={applicationData?.permitData?.contactDetails}
-          onEdit={back}
+          onEdit={handleEdit}
           continueBtnText={continueBtnText}
-          onContinue={handleSubmitForReview}
-          onAddToCart={handleAddToCart}
+          onContinue={
+            applicationStepContext === APPLICATION_STEP_CONTEXTS.APPLY
+              ? handleSubmitForReview
+              : undefined
+          }
+          onAddToCart={
+            applicationStepContext === APPLICATION_STEP_CONTEXTS.APPLY
+              ? handleAddToCart
+              : undefined
+          }
+          handleApproveButton={isQueueContext ? handleApprove : undefined}
+          handleRejectButton={isQueueContext ? handleRejectButton : undefined}
+          updateApplicationMutationPending={
+            isQueueContext ? updateApplicationMutationPending : undefined
+          }
           allConfirmed={allConfirmed}
           setAllConfirmed={setAllConfirmed}
           hasAttemptedCheckboxes={hasAttemptedSubmission}
@@ -241,16 +352,36 @@ export const ApplicationReview = ({
           vehicleWasSaved={
             applicationData?.permitData?.vehicleDetails?.saveVehicle
           }
-          vehicleConfiguration={applicationData?.permitData?.vehicleConfiguration}
+          vehicleConfiguration={
+            applicationData?.permitData?.vehicleConfiguration
+          }
           route={applicationData?.permitData?.permittedRoute}
           applicationNotes={applicationData?.permitData?.applicationNotes}
           doingBusinessAs={doingBusinessAs}
           calculatedFee={fee}
           loas={applicationData?.permitData?.loas}
           applicationRejectionHistory={applicationData?.rejectionHistory}
-          isStaffUser={isStaffUser}
+          isStaffUser={isQueueContext || isStaffUser}
         />
       </FormProvider>
+
+      {isQueueContext && showRejectApplicationModal && (
+        <RejectApplicationModal
+          showModal={showRejectApplicationModal}
+          onCancel={() => setShowRejectApplicationModal(false)}
+          onConfirm={handleReject}
+          isPending={updateApplicationMutationPending}
+        />
+      )}
+
+      {isQueueContext && showUnavailableApplicationModal && (
+        <UnavailableApplicationModal
+          showModal={showUnavailableApplicationModal}
+          onCancel={() => setShowUnavailableApplicationModal(false)}
+          onConfirm={handleCloseApplication}
+          currentClaimant={currentClaimant}
+        />
+      )}
     </div>
   );
 };
