--- frontend/src/features/queue/components/ApplicationInQueueReview.scss
@@ -1,8 +1,0 @@
-@use "../../../common/components/dashboard/Dashboard";
-
-@include Dashboard.layout-box-style(".application-in-queue-review .breadcrumb");
-
-.application-in-queue-review {
-  display: flex;
-  flex-direction: column;
-}
